feat: add web worker for 3D venue layout calculations
