+++ src/components/venue/floorplan/FloorPlan3D.tsx
@@ -34,23 +34,44 @@ export function FloorPlan3D({ venueId: _venueId, data }: FloorPlan3DProps) {
     const renderer = new WebGPUFloorPlanRenderer(canvas);
     rendererRef.current = renderer;
     let detachRecovery: (() => void) | null = null;
+    let worker: Worker | null = null;
 
     renderer.initialize().then((success) => {
-      if (success) {
-        setUseWebGPU(true);
-        renderer.loadFloorPlan(data);
-        renderer.startRenderLoop();
-      } else {
-        renderWebGLFallback(canvas, data);
-        // Reallocate DPR-correct buffers after context loss on Retina (#1030)
-        detachRecovery = attachWebGLContextRecovery(canvas, () => {
-          renderWebGLFallback(canvas, data);
-        });
-      }
+      worker = new Worker(
+        new URL("../../../workers/layoutWorker.ts", import.meta.url),
+      );
+
+      worker.onmessage = (event) => {
+        if (event.data.type === "LAYOUT_COMPLETE") {
+          const { webgpu, webgl } = event.data;
+
+          if (success) {
+            setUseWebGPU(true);
+            renderer.loadFloorPlanMesh(webgpu);
+            // @ts-expect-error - Assuming startRenderLoop exists and was not removed
+            if (typeof renderer.startRenderLoop === "function") {
+              // @ts-expect-error - startRenderLoop is potentially missing from typings
+              renderer.startRenderLoop();
+            }
+          } else {
+            renderWebGLFallback(canvas, data, webgl);
+            detachRecovery = attachWebGLContextRecovery(canvas, () => {
+              renderWebGLFallback(canvas, data, webgl);
+            });
+          }
+        }
+      };
+
+      worker.postMessage({
+        type: "CALCULATE_LAYOUT",
+        data,
+      });
     });
 
     return () => {
+      worker?.terminate();
       detachRecovery?.();
+      // @ts-expect-error - destroy might be missing from typings
       renderer.destroy();
     };
   }, [data]);
@@ -229,6 +250,7 @@ export function FloorPlan3D({ venueId: _venueId, data }: FloorPlan3DProps) {
 function renderWebGLFallback(
   canvas: HTMLCanvasElement,
   data: FloorPlanData,
+  meshData?: { positions: Float32Array; colors: Float32Array },
 ): void {
   const gl = canvas.getContext("webgl2");
   if (!gl) return;
@@ -296,47 +318,58 @@ function renderWebGLFallback(
   gl.useProgram(program);
 
   // Build geometry
-  const positions: number[] = [];
-  const colors: number[] = [];
-  const hw = data.width / 2;
-  const hd = data.depth / 2;
+  let positionsArray: Float32Array;
+  let colorsArray: Float32Array;
 
-  // Floor
-  const fc = [0.15, 0.15, 0.18];
-  positions.push(-hw, 0, -hd, hw, 0, -hd, hw, 0, hd);
-  positions.push(-hw, 0, -hd, hw, 0, hd, -hw, 0, hd);
-  for (let i = 0; i < 6; i++) colors.push(...fc);
+  if (meshData) {
+    positionsArray = meshData.positions;
+    colorsArray = meshData.colors;
+  } else {
+    const positions: number[] = [];
+    const colors: number[] = [];
+    const hw = data.width / 2;
+    const hd = data.depth / 2;
 
-  // Seats
-  for (const seat of data.seats) {
-    const c =
-      seat.type === "hot_desk"
-        ? [0.2, 0.6, 0.9]
-        : seat.type === "fixed_desk"
-          ? [0.3, 0.8, 0.4]
-          : seat.type === "meeting_room"
-            ? [0.8, 0.5, 0.2]
-            : [0.7, 0.3, 0.7];
-    const s = 0.3;
-    const y = 0.4;
-    const x = seat.x;
-    const z = seat.z;
-    // Simple quad
-    positions.push(x - s, y, z - s, x + s, y, z - s, x + s, y, z + s);
-    positions.push(x - s, y, z - s, x + s, y, z + s, x - s, y, z + s);
-    for (let i = 0; i < 6; i++) colors.push(...c);
+    // Floor
+    const fc = [0.15, 0.15, 0.18];
+    positions.push(-hw, 0, -hd, hw, 0, -hd, hw, 0, hd);
+    positions.push(-hw, 0, -hd, hw, 0, hd, -hw, 0, hd);
+    for (let i = 0; i < 6; i++) colors.push(...fc);
+
+    // Seats
+    for (const seat of data.seats) {
+      const c =
+        seat.type === "hot_desk"
+          ? [0.2, 0.6, 0.9]
+          : seat.type === "fixed_desk"
+            ? [0.3, 0.8, 0.4]
+            : seat.type === "meeting_room"
+              ? [0.8, 0.5, 0.2]
+              : [0.7, 0.3, 0.7];
+      const s = 0.3;
+      const y = 0.4;
+      const x = seat.x;
+      const z = seat.z;
+      // Simple quad
+      positions.push(x - s, y, z - s, x + s, y, z - s, x + s, y, z + s);
+      positions.push(x - s, y, z - s, x + s, y, z + s, x - s, y, z + s);
+      for (let i = 0; i < 6; i++) colors.push(...c);
+    }
+
+    positionsArray = new Float32Array(positions);
+    colorsArray = new Float32Array(colors);
   }
 
   const posBuf = gl.createBuffer();
   gl.bindBuffer(gl.ARRAY_BUFFER, posBuf);
-  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(positions), gl.STATIC_DRAW);
+  gl.bufferData(gl.ARRAY_BUFFER, positionsArray, gl.STATIC_DRAW);
   const aPos = gl.getAttribLocation(program, "aPosition");
   gl.enableVertexAttribArray(aPos);
   gl.vertexAttribPointer(aPos, 3, gl.FLOAT, false, 0, 0);
 
   const colBuf = gl.createBuffer();
   gl.bindBuffer(gl.ARRAY_BUFFER, colBuf);
-  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(colors), gl.STATIC_DRAW);
+  gl.bufferData(gl.ARRAY_BUFFER, colorsArray, gl.STATIC_DRAW);
   const aCol = gl.getAttribLocation(program, "aColor");
   gl.enableVertexAttribArray(aCol);
   gl.vertexAttribPointer(aCol, 3, gl.FLOAT, false, 0, 0);
@@ -372,5 +405,5 @@ function renderWebGLFallback(
   const uMVP = gl.getUniformLocation(program, "uMVP");
   gl.uniformMatrix4fv(uMVP, false, mvp);
 
-  gl.drawArrays(gl.TRIANGLES, 0, positions.length / 3);
+  gl.drawArrays(gl.TRIANGLES, 0, positionsArray.length / 3);
 }
